Coordenadas ajeitadas

+++ Camera/Ray.h
@@ -14,7 +14,7 @@ public:
     Ray(const Vec3& a, const Vec3& b) { A = a; B = b; }
     Vec3 origin() const       { return A; }
     Vec3 direction() const    { return B; }
-    Vec3 positionAt(float t) const { return A + t*B;}
+    Vec3 positionAt(double t) const { return A + t*B;}
 
 };
